requirements.txt
Flask==3.1.0
groq==0.17.0
langchain==0.3.17
langchain_chroma==0.2.1
langchain_community==0.3.16
langchain_huggingface==0.1.2
pymongo==4.11
python-dotenv==1.0.1
rank_bm25==0.2.2
gunicorn==20.1.0
bcrypt
Werkzeug==3.1